Removed the toString methods in ConversationPojo.java, MessagePojo.java and MessagePayloadPojo.java.
As part of the testing for NEXUS-534 and NEXUS-543, we discovered that in some situations a conversation was being loaded included all sub-objects and their subjects, turtles all the way down. After some debugging, the issue is the toString-calls to these three objects which try to return a very verbose representation, but in the process cause hibernate to load ~everything.

This fixes NEXUS-534 (and potentially NEXUS-543).