Made the null-checks on strings be StringUtils.isNotBlank-checks instead, to be safe when running on Oracle-databases.
Oracle cannot save empty strings in varchar2 fields, and instead saves NULL. This means that just because a String isn't null doesn't mean it won't violate a not-null constraint on an Oracle database. It seems the only way to get around this is to flat out make sure we never try to save an empty string to the database as part of a log message, even though this restricts functionality on other databases.

This resolves [NEXUS-613].